backend/requirements.txt
requests
langchain
langchain-groq
python-dotenv
icalendar
fastapi
uvicorn
